fix(cli): Address PR review findings
- Fix silent config failures: apply_cli_flags and apply_to_config now
  return Result<()> and propagate parsing errors
- Split config.rs (515 lines) into config/ module with types.rs,
  sections.rs, and mod.rs for better organization
- Remove unused signal-hook dependencies (using tokio::signal)
- Replace unwrap() with ? operator in version.rs JSON serialization
- Log unknown P2P events instead of silently ignoring with _ => {}
- Log error when local_peer_id() fails instead of silent ignore
- Add comprehensive tests for all config types and parsing
